src/auth/requirements.txt
jedi     
pylint 
Flask  
Flask-MySQLdb
PyJWT          